Update 5
swing animation for backhand directed to the right implemented, more animations coming in the next update. mostly everything working enough to be somewhat playable, main menu scene is broken still and there is a weird bug that teleports the ball. other than that, though, this is a huge update after update 4.